wip(TemplateUtilities.tsx): Separated fs and template utils
