+++ app/src/Common/Utilities/FsUtilities.tsx
@@ -1,40 +1,5 @@
 import { path } from '@tauri-apps/api';
-import {
-  FileEntry,
-  readDir,
-  createDir,
-  copyFile,
-  removeDir,
-  BaseDirectory,
-  exists,
-} from '@tauri-apps/api/fs';
-import Template from '../../Pages/MainPage/Types/Template';
-
-export const templatesAppdataPath = '';
-
-export const readTemplates = async (): Promise<FileEntry[]> => {
-  if (!(await exists(templatesAppdataPath, { dir: BaseDirectory.AppData }))) {
-    //create
-
-    return [];
-  }
-
-  return readDir(templatesAppdataPath, {
-    dir: BaseDirectory.AppData,
-  });
-};
-
-export const createTemplate = async (name: string, originPath: string) => {
-  if (!(await exists(templatesAppdataPath, { dir: BaseDirectory.AppData }))) {
-    // create
-  }
-
-  await createDir(name, { dir: BaseDirectory.AppData });
-  await copyFromPath(
-    originPath,
-    await path.join(await path.appDataDir(), name)
-  );
-};
+import { FileEntry, readDir, createDir, copyFile } from '@tauri-apps/api/fs';
 
 export const copyFromPath = async (origin: string, destination: string) => {
   const contents: FileEntry[] = await readDir(origin);
@@ -50,12 +15,5 @@ export const copyFromPath = async (origin: string, destination: string) => {
         await copyFile(entry.path, newPath);
       }
     }
-  });
-};
-
-export const deleteTemplate = async (template: Template) => {
-  await removeDir(template.name, {
-    dir: BaseDirectory.AppData,
-    recursive: true,
   });
 };
